reorganizing the wind package now that there are 2 wind models
