+++ smrf/distribute/wind/__init__.py
@@ -1,0 +1,1 @@
+from .wind import Wind
