+++ smrf/distribute/wind/wind.py
@@ -10,9 +10,10 @@ import pytz
 
 from smrf.distribute import image_data
 from smrf.utils import utils
+from smrf.distribute.wind.winstral import WinstralWindModel
 
 
-class wind(image_data.image_data):
+class Wind(image_data.image_data):
     """
 
     The :mod:`~smrf.distribute.wind.wind` class allows for variable specific
@@ -66,7 +67,7 @@ class wind(image_data.image_data):
     distributed using the underlying distribution methods.
 
     Args:
-        windConfig: The [wind] section of the configuration file
+        self.config: The [wind] section of the configuration file
 
     Attributes:
         config: configuration from [vapor_pressure] section
@@ -111,20 +112,25 @@ class wind(image_data.image_data):
     # be written during main distribute loop
     post_process_variables = {}
 
-    def __init__(self, windConfig, distribute_drifts, wholeConfig, tempDir=None):
+    def __init__(self, config):
 
         # extend the base class
         image_data.image_data.__init__(self, self.variable)
         self._logger = logging.getLogger(__name__)
 
         # check and assign the configuration
-        self.getConfig(windConfig)
+        self.smrf_config = config
+        self.getConfig(config['wind'])
 
-        if (tempDir is None) | (tempDir == 'WORKDIR'):
-            tempDir = os.environ['WORKDIR']
-        self.tempDir = tempDir
+        # if (tempDir is None) | (tempDir == 'WORKDIR'):
+        #     tempDir = os.environ['WORKDIR']
+        # self.tempDir = tempDir
 
-        if windConfig['distribution'] == 'grid':
+        distribute_drifts = False
+        if self.smrf_config["precip"]["precip_rescaling_model"] == "winstral":
+            distribute_drifts = True
+
+        if self.config['distribution'] == 'grid':
             self.gridded = True
             self.distribute_drifts = False
 
@@ -136,44 +142,13 @@ class wind(image_data.image_data):
                 self.wind_ninja_tz = pytz.timezone(
                     self.config['wind_ninja_tz'].title())
 
-            self.start_date = pd.to_datetime(wholeConfig['time']['start_date'])
-            self.grid_data = wholeConfig['gridded']['data_type']
+            self.start_date = pd.to_datetime(
+                self.smrf_config['time']['start_date'])
+            self.grid_data = self.smrf_config['gridded']['data_type']
 
         else:
-            # open the maxus netCDF
-            self._maxus_file = nc.Dataset(self.config['maxus_netcdf'], 'r')
-            self.maxus = self._maxus_file.variables['maxus'][:]
-            self.maxus_direction = self._maxus_file.variables['direction'][:]
-            self._maxus_file.close()
-
-            # Maxus must be the the same size as the topo.
-            topo_nc = nc.Dataset(wholeConfig['topo']['filename'], 'r')
-            t_shape = topo_nc.variables['dem'].shape
-            topo_nc.close()
-
-            if t_shape != self.maxus[0].shape:
-                raise IOError("\nMaxus file must be generated using the topo to"
-                              " be valid. Maxus netcdf shape = {} and topo"
-                              " netcdf shape = {}".format(t_shape,
-                                                          self.maxus.shape))
-
-            self._logger.debug('Read data from {}'
-                               .format(self.config['maxus_netcdf']))
-
-            # get the veg values
-            matching = [s for s in self.config.keys() if "veg_" in s]
-            v = {}
-            for m in matching:
-                ms = m.split('_')
-                if type(self.config[m]) == list:
-                    v[ms[1]] = float(self.config[m][0])
-                else:
-                    v[ms[1]] = float(self.config[m])
-
-            self.veg = v
-
-            # whether or not we will use this data to redistribute precip
-            self.distribute_drifts = distribute_drifts
+            self.wind_model = WinstralWindModel(
+                self.smrf_config, distribute_drifts)
 
         self._logger.debug('Created distribute.wind')
 
@@ -192,41 +167,15 @@ class wind(image_data.image_data):
         """
 
         self._logger.debug('Initializing distribute.wind')
-        self._initialize(topo, data.metadata)
-
-        # meshgrid points
-        self.X = topo.X
-        self.Y = topo.Y
 
         if not self.gridded:
-            self.veg_type = topo.veg_type
+            self.wind_model.initialize(topo, data)
 
-            # get the enhancements for the stations
-            if 'enhancement' not in self.metadata.columns:
-                self.metadata['enhancement'] = \
-                    float(self.config['station_default'])
-
-                for m in self.metadata.index:
-                    sta_e = m.lower()
-                    if sta_e in self.config:
-                        if type(self.config[sta_e]) == list:
-                            enhancement = self.config[sta_e][0]
-                        else:
-                            enhancement = self.config[sta_e]
-
-                        self.metadata.loc[m, 'enhancement'] = \
-                            float(enhancement)
         else:
             self.flatwind = None
 
             if self.config['wind_ninja_dir'] is not None:
                 self.initialize_wind_ninja(topo)
-
-        if not self.distribute_drifts:
-            # we have to pass these to precip, so make them none if we won't use them
-            self.dir_round_cell = None
-            self.cellmaxus = None
-
 
     def initialize_wind_ninja(self, topo):
 
@@ -235,7 +184,7 @@ class wind(image_data.image_data):
         # set roughness that was used in WindNinja simulation
         # WindNinja uses 0.01m for grass, 0.43 for shrubs, and 1.0 for forest
         self.wn_roughness = float(self.config['wind_ninja_roughness']) * \
-                            np.ones_like(topo.dem)
+            np.ones_like(topo.dem)
 
         # get our effective veg surface roughness
         # to use in log law scaling of WindNinja data
@@ -251,15 +200,16 @@ class wind(image_data.image_data):
 
         # precalculate scale arrays so we don't do it every timestep
         self.ln_wind_scale = np.log((self.veg_roughness + self.wind_height) / self.veg_roughness) / \
-                                np.log((self.wn_roughness + self.wind_height) / self.wn_roughness)
+            np.log((self.wn_roughness + self.wind_height) / self.wn_roughness)
 
         ###### do this first to speedup the interpolation later ####
         # find vertices and weights to speedup interpolation fro ascii file
         fmt_d = '%Y%m%d'
         fp_vel = glob.glob(os.path.join(self.wind_ninja_dir,
-                                'data{}'.format(self.start_date.strftime(fmt_d)),
-                                'wind_ninja_data',
-                                '*{}m_vel.asc'.format(self.wind_ninja_dxy)))[0]
+                                        'data{}'.format(
+                                            self.start_date.strftime(fmt_d)),
+                                        'wind_ninja_data',
+                                        '*{}m_vel.asc'.format(self.wind_ninja_dxy)))[0]
 
         # get wind ninja topo stats
         ts2 = utils.get_asc_stats(fp_vel)
@@ -272,14 +222,14 @@ class wind(image_data.image_data):
         self.wn_mx = xwint
         self.wn_my = ywint
 
-        xy = np.zeros([XW.shape[0]*XW.shape[1],2])
-        xy[:,1] = ywint
-        xy[:,0] = xwint
-        uv = np.zeros([self.X.shape[0]*self.X.shape[1],2])
-        uv[:,1] = self.Y.flatten()
-        uv[:,0] = self.X.flatten()
+        xy = np.zeros([XW.shape[0]*XW.shape[1], 2])
+        xy[:, 1] = ywint
+        xy[:, 0] = xwint
+        uv = np.zeros([self.X.shape[0]*self.X.shape[1], 2])
+        uv[:, 1] = self.Y.flatten()
+        uv[:, 0] = self.X.flatten()
 
-        self.vtx, self.wts = utils.interp_weights(xy, uv,d=2)
+        self.vtx, self.wts = utils.interp_weights(xy, uv, d=2)
         ###### end do this first to speedup the interpolation later ####
 
     def distribute(self, data_speed, data_direction, t):
@@ -309,8 +259,8 @@ class wind(image_data.image_data):
         self._logger.debug('{} Distributing wind_direction and wind_speed'
                            .format(data_speed.name))
 
-        data_speed = data_speed[self.stations]
-        data_direction = data_direction[self.stations]
+        data_speed = data_speed[self.wind_model.stations]
+        data_direction = data_direction[self.wind_model.stations]
 
         if self.gridded:
             # check if running with windninja
@@ -344,21 +294,7 @@ class wind(image_data.image_data):
                                                 self.max)
 
         else:
-            # calculate the maxus at each site
-            self.stationMaxus(data_speed, data_direction)
-
-            # distribute the flatwind
-            self._distribute(self.flatwind_point,
-                             other_attribute='flatwind')
-
-            # distribute u_direction and v_direction
-            self._distribute(self.u_direction,
-                             other_attribute='u_direction_distributed')
-            self._distribute(self.v_direction,
-                             other_attribute='v_direction_distributed')
-
-            # Calculate simulated wind speed at each cell from flatwind
-            self.simulateWind(data_speed)
+            self.wind_model.distribute(data_speed, data_direction)
 
     def distribute_thread(self, queue, data_speed, data_direction):
         """
@@ -377,192 +313,11 @@ class wind(image_data.image_data):
 
             self.distribute(data_speed.loc[t], data_direction.loc[t], t)
 
-            queue['wind_speed'].put([t, self.wind_speed])
-            queue['wind_direction'].put([t, self.wind_direction])
-            queue['flatwind'].put([t, self.flatwind])
-            queue['cellmaxus'].put([t, self.cellmaxus])
-            queue['dir_round_cell'].put([t, self.dir_round_cell])
-
-    def simulateWind(self, data_speed):
-        """
-        Calculate the simulated wind speed at each cell from flatwind and the
-        distributed directions. Each cell's maxus value is pulled from the
-        maxus library based on the distributed wind direction. The cell's maxus
-        is further adjusted based on the vegetation type and the factors
-        provided in the [wind] section of the configuration file.
-
-        Args:
-            data_speed: Pandas dataframe for a single time step of wind speed
-                to make the pixel locations same as the measured values
-        """
-
-        # combine u and v to azimuth
-        az = np.arctan2(self.u_direction_distributed,
-                        self.v_direction_distributed)*180/np.pi
-        az[az < 0] = az[az < 0] + 360
-
-        dir_round_cell = np.ceil((az - self.nstep/2) / self.nstep) * self.nstep
-        dir_round_cell[dir_round_cell <
-                       0] = dir_round_cell[dir_round_cell < 0] + 360
-        dir_round_cell[dir_round_cell == -0] = 0
-        dir_round_cell[dir_round_cell == 360] = 0
-
-        cellmaxus = np.zeros(dir_round_cell.shape)
-        cellwind = np.zeros(dir_round_cell.shape)
-
-        dir_unique = np.unique(dir_round_cell)
-        for d in dir_unique:
-            # find all values for matching direction
-            ind = dir_round_cell == d
-            i = np.argwhere(self.maxus_direction == d)[0][0]
-            cellmaxus[ind] = self.maxus[i][ind]
-
-        # correct for veg
-        dynamic_mask = np.ones(cellmaxus.shape)
-        for k, v in self.veg.items():
-            # Adjust veg types that were specified by the user
-            if k != 'default':
-                ind = self.veg_type == int(k)
-                dynamic_mask[ind] = 0
-                cellmaxus[ind] += v
-
-        # Apply the veg default to those that weren't messed with
-        if self.veg['default'] != 0:
-            cellmaxus[dynamic_mask == 1] += self.veg['default']
-
-        # correct unreasonable values
-        cellmaxus[cellmaxus > 32] = 32
-        cellmaxus[cellmaxus < -32] = -32
-
-        # determine wind
-        factor = float(self.config['reduction_factor'])
-        ind = cellmaxus < -30.10
-        cellwind[ind] = factor * self.flatwind[ind] * 4.211
-
-        ind = (cellmaxus > -30.10) & (cellmaxus < -21.3)
-        c = np.abs(cellmaxus[ind])
-        cellwind[ind] = factor * self.flatwind[ind] * \
-            (1.756507 - 0.1678945 * c + 0.01927844 * np.power(c, 2) -
-             0.0003651592 * np.power(c, 3))
-
-        ind = (cellmaxus > -21.3) & (cellmaxus < 0)
-        c = np.abs(cellmaxus[ind])
-        cellwind[ind] = factor * self.flatwind[ind] * \
-            (1.0 + 0.1031717 * c - 0.008003561 * np.power(c, 2) +
-             0.0003996581 * np.power(c, 3))
-
-        ind = cellmaxus > 30.10
-        cellwind[ind] = self.flatwind[ind] / 4.211
-
-        ind = (cellmaxus < 30.10) & (cellmaxus > 21.3)
-        c = cellmaxus[ind]
-        cellwind[ind] = self.flatwind[ind] / \
-            (1.756507 - 0.1678945 * c + 0.01927844 * np.power(c, 2) -
-             0.0003651592 * np.power(c, 3))
-
-        ind = (cellmaxus < 21.3) & (cellmaxus >= 0)
-        c = cellmaxus[ind]
-        cellwind[ind] = self.flatwind[ind] / \
-            (1.0 + 0.1031717 * c - 0.008003561 * np.power(c, 2) +
-             0.0003996581 * np.power(c, 3))
-
-        # Convert from 3m to 5m wind speed
-        cellwind *= 1.07985
-
-        # preseve the measured values
-        cellwind[self.metadata.yi, self.metadata.xi] = data_speed
-
-        # check for NaN
-        nans, x = utils.nan_helper(cellwind)
-
-        if np.sum(nans) > 0:
-            cellwind[nans] = np.interp(x(nans), x(~nans), cellwind[~nans])
-
-        self.wind_speed = utils.set_min_max(cellwind, self.min, self.max)
-        self.wind_direction = az
-        self.cellmaxus = cellmaxus
-        self.dir_round_cell = dir_round_cell
-
-    def stationMaxus(self, data_speed, data_direction):
-        """
-        Determine the maxus value at the station given the wind direction.
-        Can specify the enhancemet for each station or use the default, along
-        with whether or not the station is on a peak which will ensure that
-        the station cannot be sheltered. The station enhancement and peak
-        stations are specified in the [wind] section of the configuration
-        file. Calculates the following for each station:
-
-        * :py:attr:`flatwind`
-        * :py:attr:`u_direction`
-        * :py:attr:`v_direction`
-
-        Args:
-            data_speed: wind_speed data frame for single time step
-            data_direction: wind_direction data frame for single time step
-
-        """
-
-        # ----------------------------------------
-        # Get data and site maxus value
-        flatwind = data_speed.copy()
-
-        # number of bins that the maxus library was calculated for
-        self.nbins = len(self.maxus_direction)
-        self.nstep = 360/self.nbins
-
-        for m in self.metadata.index:
-            # pixel locations
-            xi = self.metadata.loc[m, 'xi']
-            yi = self.metadata.loc[m, 'yi']
-            e = self.metadata.loc[m, 'enhancement']
-
-            # maxus value at the station
-            if not pd.isnull(data_direction[m]):
-                if self.config['station_peak'] != None:
-                    if m.upper() in self.config['station_peak']:
-                        val_maxus = np.min(self.maxus[:, yi, xi] + e)
-
-                else:
-                    idx = int(np.ceil((data_direction[m] - self.nstep/2) /
-                                      self.nstep) * self.nstep)
-                    if idx == 360:
-                        idx = 0  # special case when 360=0
-                    ind = self.maxus_direction == idx
-
-                    val_maxus = self.maxus[ind, yi, xi] + e
-
-                # correct unreasonable values
-                if val_maxus > 35:
-                    val_maxus = 35
-                if val_maxus < -35:
-                    val_maxus = -35
-
-                ma = np.abs(val_maxus)
-
-                # Lapse all measurements to flat terrain (i.e. maxus = 0)
-                if (ma > 21.3 and ma < 30.0):
-                    expVal = 1.756507 - 0.1678945 * ma + \
-                        0.01927844 * np.power(ma, 2) - \
-                        0.0003651592 * np.power(ma, 3)
-                elif (ma >= 30.0):
-                    expVal = 4.21
-                else:
-                    expVal = 1.0 + 0.1031717 * (ma) - \
-                        0.008003561 * np.power(ma, 2) + \
-                        0.0003996581 * np.power(ma, 3)
-
-                if val_maxus > 0:
-                    flatwind.loc[m] = data_speed[m] * expVal
-                else:
-                    flatwind.loc[m] = data_speed[m] / expVal
-            else:
-                flatwind.loc[m] = np.NaN
-
-        self.flatwind_point = flatwind
-
-        # wind direction components at the station
-        self.u_direction = np.sin(data_direction * np.pi/180)    # u
-        self.v_direction = np.cos(data_direction * np.pi/180)    # v
+            queue['wind_speed'].put([t, self.wind_model.wind_speed])
+            queue['wind_direction'].put([t, self.wind_model.wind_direction])
+            queue['flatwind'].put([t, self.wind_model.flatwind])
+            queue['cellmaxus'].put([t, self.wind_model.cellmaxus])
+            queue['dir_round_cell'].put([t, self.wind_model.dir_round_cell])
 
     def convert_wind_ninja(self, t):
         """
